requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
pyjwt
python-multipart
python-dotenv
openai-whisper
ffmpeg-python
